edit gradle test execution to be sequential
Problem is that on parallel execution one of controller tests can't
startup local server, because it's already in use. Yes, i've tried
to change tomcat's default port and also to change spring boot's web
env to random port. I've checked that ports are indeed random, rest
assured was aware of them yet still problem was there. Sorry for
inconvenience. This is temporal solution, but i've spend too
much time for this for now.